% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dx.R
\name{check_job}
\alias{check_job}
\title{Watches a job given a job id}
\usage{
check_job(job_id)
}
\arguments{
\item{job_id}{\itemize{
\item job id in the format of \code{job-XXXXXX}.
}}
}
\value{
file_id - If a CSV file was generated in the job, otherwise NULL. Prints alert
to screen of job status and job_id for file generated.
}
\description{
Watches a job given a job id
}
\examples{
\dontrun{
ds_id <- find_dataset_id()
fields <- fields <- c("participant.eid", "participant.p31", "participant.p41202")
job_id <- launch_table_exporter(ds_id, fields)

check_job(job_id)

}

}
