% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project.R
\name{decode_df_db}
\alias{decode_df_db}
\title{Main Function to Decode Fields from Codings using Duck DB}
\usage{
decode_df_db(df, coding)
}
\arguments{
\item{coding}{}
}
\description{
Main Function to Decode Fields from Codings using Duck DB
}
\examples{
data(coding_dict)
data(cohort)
data(data_dict)

cdata <- merge_coding_data_dict(coding_dict, data_dict)

cohort |>
  decode_df_db(cdata)

}
