% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project.R
\name{decode_multi}
\alias{decode_multi}
\title{Decodes Multi Category variables}
\usage{
decode_multi(cohort, coding)
}
\arguments{
\item{coding}{}
}
\value{
Labeled data frame where list columns are decoded
}
\description{
Decodes Multi Category variables
}
\examples{
data(coding_dict)
data(cohort)
data(data_dict)
}
