% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project.R
\name{decode_multi_db}
\alias{decode_multi_db}
\title{Decodes Multi Category variables}
\usage{
decode_multi_db(cohort, coding)
}
\arguments{
\item{coding}{}
}
\value{
Labeled data frame where multi category columns are decoded
}
\description{
Decodes Multi Category variables
}
\examples{
data(coding_dict)
data(cohort)
data(data_dict)

cdata <- merge_coding_data_dict(coding_dict, data_dict)

cohort |>
  decode_multi(cdata)

}
