% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project.R
\name{decode_multi_large_df}
\alias{decode_multi_large_df}
\title{Decodes large columns of a large data frame by splitting it}
\usage{
decode_multi_large_df(df, coding, df_size = 2000)
}
\arguments{
\item{df}{\itemize{
\item data.frame to process
}}

\item{coding}{\itemize{
\item coding data frame
}}

\item{df_size}{\itemize{
\item size of the smaller data frames in number of rows.
}}
}
\value{
decoded data frame in terms of columns
}
\description{
Given an extremely large data frame to decode (such as those of UK Biobank,
over 500K rows), split the data into a list of smaller data frames, decode
each smaller data frame, and then merge list into a larger data frame.
}
\details{
This method is not necessarily faster than merging the entire data frame,
but it shows a progress bar.
}
