% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project.R
\name{decode_multi_purrr}
\alias{decode_multi_purrr}
\title{Decodes Multi Category variables}
\usage{
decode_multi_purrr(cohort, coding)
}
\arguments{
\item{cohort}{\itemize{
\item cohort or dataset in data.frame format
}}

\item{coding}{}
}
\value{
Labeled data frame where multi category columns are decoded
}
\description{
For a smaller data frame, decodes the columns that have a list of values.
}
\details{
For a larger data frame, such as from UK Biobank, we recommend you use
\code{decode_multi_large_df()}, which will split your data frame into
smaller ones for decoding.
}
\examples{
data(coding_dict)
data(cohort)
data(data_dict)

cdata <- merge_coding_data_dict(coding_dict, data_dict)

cohort |>
  decode_multi_purrr(cdata)

}
