% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project.R
\name{decode_single}
\alias{decode_single}
\title{Decodes variables with single coded entries}
\usage{
decode_single(cohort, coding)
}
\arguments{
\item{coding}{}
}
\value{
data.frame with single coded columns decoded
}
\description{
Decodes variables with single coded entries
}
