% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project.R
\name{explore_field_list}
\alias{explore_field_list}
\title{Given a data dictionary, displays a searachable table in a quarto document or Jupyter Notebook}
\usage{
explore_field_list(ds_id, path = ".")
}
\arguments{
\item{coding_data_dict}{\itemize{
\item a coding/data dictionary created by merge_coding_data_dict or
by build_data_dict()
}}
}
\value{
none - a reactable table will be created in the document, with a searchable window.
}
\description{
This function leverages reactable to give a searchable table of fields. It's most useful
in a Jupyter Notebook or RMarkdown/Quarto Document
}
\examples{
if(interactive()){
  data(coding_dict)
  data(cohort)
  data(data_dict)

  cdata <- merge_coding_data_dict(coding_dict, data_dict)

  explore_field_list(cdata)
}
}
