% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dx.R
\name{extract_data}
\alias{extract_data}
\title{Given a dataset id and list of fields, extract data}
\usage{
extract_data(dataset_id, field_list)
}
\arguments{
\item{dataset_id}{\itemize{
\item dataset id in the form project-xxxx:record-yyyy- you can use get_dataset_id() to find this.
}}

\item{field_list}{\itemize{
\item vector/list of fields. Note that you can choose to return eids below
}}

\item{return_eids}{\itemize{
\item Whether to return eids
}}
}
\value{
Alert to dataset location
}
\description{
Given a dataset id and list of fields, extract data
}
\examples{

\dontrun{
fields <- c("participant.eid", "participant.p31", "participant.p41202")
ds_id <- get_dataset_id()
extract_data(ds_id, field_list=fields)
}
}
