% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dx.R
\name{find_all_cohorts}
\alias{find_all_cohorts}
\title{Finds all cohorts and their ids in the current project}
\usage{
find_all_cohorts()
}
\value{
\code{data.frame} listing all cohorts in the project.
Cohort ID, name, project_id, project_record ID, and the
linked dataset ID is returned
}
\description{
Finds all cohorts and their ids in the current project
}
\examples{

\dontrun{
find_all_cohorts()
}
}
