% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dx.R
\name{find_all_datasets}
\alias{find_all_datasets}
\title{Find all datasets in the current project}
\usage{
find_all_datasets()
}
\value{
data.frame of all datasets in the project
}
\description{
Find all datasets in the current project
}
