% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dx.R
\name{find_all_jobs}
\alias{find_all_jobs}
\title{Finds all jobs and executions in current project}
\usage{
find_all_jobs()
}
\value{
\itemize{
\item data.frame with the following columns:
``
}
}
\description{
Finds all jobs and executions in current project
}
