% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dx.R
\name{find_dataset_id}
\alias{find_dataset_id}
\title{Finds the current dataset id in the project}
\usage{
find_dataset_id()
}
\value{
\itemize{
\item dataset id in the form project-xxxx:record-yyyy
}
}
\description{
Finds the current dataset id in the project
}
\examples{
\dontrun{
find_dataset_id()
}
}
