% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dx.R
\name{get_coding_table}
\alias{get_coding_table}
\title{Given a dataset id, builds a coding table to decode raw data}
\usage{
get_coding_table(ds_id, path = ".")
}
\arguments{
\item{ds_id}{\itemize{
\item dataset id, with the format \code{project-XXXXX:record-YYYY}
}}

\item{path}{\itemize{
\item path in current project
}}
}
\value{
data.frame of codings merged with data dictionary for dataset/
cohort
}
\description{
Given a dataset id, builds a coding table to decode raw data
}
\examples{
\dontrun{
ds_id <- find_dataset_id()
get_dictionaries(ds_id)
codings <- get_coding_table(ds_id)
head(codings)
}
}
