% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dx.R
\name{get_dictionaries}
\alias{get_dictionaries}
\title{Extracts the dictionary files for a dataset id into current project}
\usage{
get_dictionaries(dataset_id = NULL)
}
\arguments{
\item{dataset_id}{}
}
\value{
side effect is downloading the \code{data.dictionary.csv}, \code{codings.csv},
and \code{entity.csv} file for the dataset.
}
\description{
Extracts the dictionary files for a dataset id into current project
}
\examples{
\dontrun{
ds_id <- get_dataset_id()
get_dictionaries(ds_id)
}
}
