% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dx.R
\name{get_name_from_full_id}
\alias{get_name_from_full_id}
\title{Given a dataset id, finds the dataset name}
\usage{
get_name_from_full_id(id)
}
\arguments{
\item{id}{}
}
\value{
ds_name - the name of the dataset
}
\description{
Given a dataset id, finds the dataset name
}
