% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dx.R
\name{install_dxpy}
\alias{install_dxpy}
\title{Installs the dx-toolkit into a virtualenv}
\usage{
install_dxpy(method = "auto", conda = "auto")
}
\value{
no-value. Side effect is that dxpy and pandas are installed
into \code{r-reticulate} environment.
}
\description{
Not needed for running in JupyterLab/RStudio on UKB RAP or JupyterLab on
core platform, as the dx-toolkit is installed on these apps.
}
\details{
Useful for using xvhelper on your own machine.
}
\examples{

\dontrun{
install_dxpy()
}
}
