% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dx.R
\name{launch_table_exporter}
\alias{launch_table_exporter}
\title{Runs Table Exporter}
\usage{
launch_table_exporter(ds_id, field_list, ...)
}
\arguments{
\item{...}{}
}
\value{
job_id Job ID of launched job
}
\description{
Runs Table Exporter
}
\examples{

\dontrun{
ds_id <- find_dataset_id()
fields <- fields <- c("participant.eid", "participant.p31", "participant.p41202")
job_id <- launch_table_exporter(ds_id, fields)

# Watch our job
check_job(job_id)

# Terminate our job if necessary
terminate_job(job_id)

}
}
