% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dx.R
\name{list_fields}
\alias{list_fields}
\title{Title}
\usage{
list_fields(dataset_id = NULL)
}
\arguments{
\item{dataset_id}{\itemize{
\item ID of the dataset, in \code{project-XXXX:record-YYYY} format
}}
}
\description{
Title
}
