% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project.R
\name{map_categories_multi_column}
\alias{map_categories_multi_column}
\title{Title}
\usage{
map_categories_multi_column(column_name, df_to_convert, coded_col_df)
}
\arguments{
\item{coded_col_df}{}
}
\value{

}
\description{
Title
}
