% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project.R
\name{merge_coding_data_dict}
\alias{merge_coding_data_dict}
\title{Title}
\usage{
merge_coding_data_dict(coding_dict, data_dict)
}
\arguments{
\item{data_dict}{}
}
\description{
Title
}
\examples{
data(coding_dict)
data(cohort)
data(data_dict)

cdata <- merge_coding_data_dict(coding_dict, data_dict)
head(cdata)
}
