% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project.R
\name{search_field_list}
\alias{search_field_list}
\title{Title}
\usage{
search_field_list(dict)
}
\arguments{
\item{dict}{}
}
\value{

}
\description{
Title
}
\examples{
data(coding_dict)
data(cohort)
data(data_dict)

cdata <- merge_coding_data_dict(coding_dict, data_dict)

search_field_list(cdata)

}
