% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dx.R
\name{terminate_job}
\alias{terminate_job}
\title{Terminates a job using job id}
\usage{
terminate_job(job_id)
}
\arguments{
\item{job_id}{}
}
\value{
Nothing - alert that job has been terminated.
}
\description{
Terminates a job using job id
}
