% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project.R
\name{decode_column_names}
\alias{decode_column_names}
\title{Returns cleaned field titles for a cohort dataset}
\usage{
decode_column_names(cohort, coded_col_df)
}
\arguments{
\item{coded_col_df}{}
}
\value{

}
\description{
Returns cleaned field titles for a cohort dataset
}
