% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project.R
\name{decode_single}
\alias{decode_single}
\title{Decodes variables with single coded entries}
\usage{
decode_single(cohort, coding)
}
\arguments{
\item{cohort}{\itemize{
\item A cohort or dataset extracted using \verb{dx extract data}
}}

\item{coding}{\itemize{
\item Combined coding/data dictionary generated from \code{merge_coding_data_dict}
}}
}
\value{
data.frame with single coded columns decoded
}
\description{
Decodes variables with single coded entries
}
