% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project.R
\name{explore_field_list}
\alias{explore_field_list}
\title{Given a data dictionary, displays a searachable table in a quarto document or Jupyter Notebook}
\usage{
explore_field_list(dict)
}
\arguments{
\item{dict}{}
}
\value{
none - a reactable table will be created in the document, with a searchable window.
}
\description{
Given a data dictionary, displays a searachable table in a quarto document or Jupyter Notebook
}
\examples{
data(coding_dict)
data(cohort)
data(data_dict)

cdata <- merge_coding_data_dict(coding_dict, data_dict)

search_field_list(cdata)

}
